app/src/main/java/piuk/blockchain/android/ui/dashboard/announcements/rule/BitpayAnnouncement.kt
package piuk.blockchain.android.ui.dashboard.announcements.rule

import android.support.annotation.VisibleForTesting
import com.blockchain.preferences.WalletStatus
import io.reactivex.Single
import piuk.blockchain.android.R
import piuk.blockchain.android.ui.dashboard.announcements.AnnouncementCard
import piuk.blockchain.android.ui.dashboard.announcements.AnnouncementHost
import piuk.blockchain.android.ui.dashboard.announcements.AnnouncementRule
import piuk.blockchain.android.ui.dashboard.announcements.DismissRecorder
import piuk.blockchain.android.ui.dashboard.announcements.DismissRule

class BitpayAnnouncement(
    dismissRecorder: DismissRecorder,
    val walletStatus: WalletStatus
) : AnnouncementRule(dismissRecorder) {

    override val dismissKey = DISMISS_KEY

    override fun shouldShow(): Single<Boolean> {
        if (dismissEntry.isDismissed) {
            return Single.just(false)
        }

        return Single.just(!walletStatus.hasMadeBitPayTransaction)
    }

    override fun show(host: AnnouncementHost) {
        host.showAnnouncementCard(
            card = AnnouncementCard(
                name = name,
                dismissRule = DismissRule.CardOneTime,
                dismissEntry = dismissEntry,
                bodyText = R.string.bitpay_announcement_body,
                iconImage = R.drawable.ic_bitpay_logo,
                dismissFunction = {
                    host.dismissAnnouncementCard(dismissEntry.prefsKey)
                }
            )
        )
    }

    override val name = "bitpay"

    companion object {
        @VisibleForTesting(otherwise = VisibleForTesting.PRIVATE)
        const val DISMISS_KEY = "BitpayAnnouncement_DISMISSED"
    }
}
